Updating the rest of the actions.
